refactor: MediaQuery 사이즈 훅으로 분리 & 상수 추가로 언제나 모바일로 보여주도록 하기
