+++ src/components/pages/performances/PerformancesSearchInput.tsx
@@ -1,13 +1,13 @@
 'use client';
 import React, { useState } from 'react';
-import { useMediaQuery } from 'react-responsive';
 import SearchInput from '@/components/common/SearchInput/SearchInput';
+import { useIsMobile } from '@/hooks';
 import useQueryParam from '@/hooks/useQueryParam/useQueryParam';
 
 const PerformancesSearchInput = () => {
   const { setMultipleQueryParams, getQueryParam } = useQueryParam();
   const [keyword, setKeyword] = useState(getQueryParam('keyword') || '');
-  const isMobile = useMediaQuery({ maxWidth: 768 });
+  const isMobile = useIsMobile();
 
   const handleSubmit = () => {
     setMultipleQueryParams({
