+++ src/constants/index.ts
@@ -1,4 +1,5 @@
+import { IS_MOBILE_IN_DESKTOP } from './layout';
 import { LocationLabels } from './locationLabels';
 import { LocationValues } from './locationValues';
 
-export { LocationLabels, LocationValues };
+export { LocationLabels, LocationValues, IS_MOBILE_IN_DESKTOP };
