+++ src/constants/layout.ts
@@ -1,0 +1,2 @@
+const IS_MOBILE_IN_DESKTOP = true;
+export { IS_MOBILE_IN_DESKTOP };
