+++ src/hooks/index.ts
@@ -1,7 +1,8 @@
 export { default as useStyles } from './useStyles/useStyles';
 export { default as useModalController } from './useModalController/useModalController';
-export { default as useIsMobile } from './useCheckIsMobile/useCheckIsMobile';
 export { usePatchPerformanceLiked } from './performanceHooks/performanceHooks';
+export { default as useIsMobile } from './sizeHook/useIsMobile';
+export { default as useIsDesktop } from './sizeHook/useIsDesktop';
 export { default as useClickOutside } from './useClickOutside/useClickOutside';
 export { default as useQueryParam } from './useQueryParam/useQueryParam';
